==> dataProcessingAlgorithms/step1.MergeOriginCsvsToOneCsv.ipynb <==
import os, glob
import pandas as pd

path = ["../datasets/origin/12.12.-splits/train/","../datasets/origin/12.12.-splits/test/","../datasets/origin/12.12.-splits/val/"]
name = ["train","test","val"]

for i in range (3):
    all_df = glob.glob(os.path.join(path[i], "df_*.csv"))
    df_from_each_file = (pd.read_csv(f, sep=',').loc[:,'gameId':] for f in all_df)
    df_merged  = pd.concat(df_from_each_file, ignore_index=True)
    df_merged.to_csv("../datasets/merged_data/"+name[i]+".csv", index= False)
