==> dataProcessingAlgorithms/step2.ChampionJsonToDF.ipynb <==
import pandas as pd
from tqdm import tqdm
import os
import json
from glom import glom

df = pd.read_json('../datasets/merged_data/champion.json')
df= df.T

df['key'] = pd.to_numeric(df['key'], errors='ignore')
df.sort_values("key" , ascending=True,inplace=True)
df.head(5)

df_champion = df[['name', 'key']]
df_champion.columns=['name', 'id']
df_champion


df_champion.to_csv( "champion.csv", index= False)

df_champion['cj'] = 0
df_champion2 = df_champion.copy()

combo = df_champion.merge(df_champion2, on='cj', how='outer')

combo

combo.drop(['cj'], axis=1, inplace=True)
combo = combo.loc[combo['id_y'] > combo['id_x']]
combo.columns = ['champion1','id1','champion2','id2']
combo['combo'] = 0

combo

combo.to_csv( "combo.csv", index= False)
