==> dataProcessingAlgorithms/step4.SplitByTime.ipynb <==
import pandas as pd
from tqdm import tqdm
import os

data_train = pd.read_csv("../datasets/processed_data/train.csv")
data_test= pd.read_csv("../datasets/processed_data/test.csv")
data_val = pd.read_csv("../datasets/processed_data/val.csv")

# Minute segments of 20 and 30 minutes
min_split1 = 20
min_split2 = 30

time_split_in_sec =[60*min_split1,60*min_split2]
time_split_in_sec 

def split_by_time(df):
    df_pre = df[df['gameDuration'] < time_split_in_sec[0]]
    df_mid = df.loc[(df['gameDuration'] > time_split_in_sec[0]) & (df['gameDuration'] < time_split_in_sec[1])]
    df_late = df[df['gameDuration'] > time_split_in_sec[1]]
    print ("pre: " + str(df_pre.shape[0]) + "\n" 
    + "mid: " + str(df_mid.shape[0]) +"\n" 
    + "late: " +str(df_late.shape[0]))
    return df_pre,df_mid,df_late

df_pre,df_mid,df_late = split_by_time(data_train)
df_pre.to_csv("../datasets/train_pre.csv", index=False)
df_mid.to_csv("../datasets/train_mid.csv", index=False)
df_late.to_csv("../datasets/train_late.csv", index=False)

df_pre,df_mid,df_late = split_by_time(data_val)
df_pre.to_csv("../datasets/val_pre.csv", index=False)
df_mid.to_csv("../datasets/val_mid.csv", index=False)
df_late.to_csv("../datasets/val_late.csv", index=False)
