==> dataProcessingAlgorithms/step5.1.ChampionCounterScores.ipynb <==
import pandas as pd
from tqdm import tqdm
import numpy as np

pre = pd.read_csv("../datasets/train_pre.csv")
mid = pd.read_csv("../datasets/train_mid.csv")
late = pd.read_csv("../datasets/train_late.csv")


# def ChampCounterRelations(df):
#     top_counter = df[['Player_1_pick','Player_1_gold','Player_6_pick','Player_6_gold']].groupby(['Player_1_pick','Player_6_pick']).sum().reset_index()
#     top_counter['counter_Score'] = top_counter.apply(lambda x: x['Player_1_gold']-x['Player_6_gold'], axis=1)
#     jug_counter = df[['Player_2_pick','Player_2_gold','Player_7_pick','Player_7_gold']].groupby(['Player_2_pick','Player_7_pick']).mean().reset_index()
#     jug_counter['counter_Score'] = jug_counter.apply(lambda x: x['Player_2_gold']-x['Player_7_gold'], axis=1)
#     mid_counter = df[['Player_3_pick','Player_3_gold','Player_8_pick','Player_8_gold']].groupby(['Player_3_pick','Player_8_pick']).mean().reset_index()
#     mid_counter['counter_Score'] = mid_counter.apply(lambda x: x['Player_3_gold']-x['Player_8_gold'], axis=1)
#     bot_counter = df[['Player_4_pick','Player_4_gold','Player_9_pick','Player_9_gold']].groupby(['Player_4_pick','Player_9_pick']).mean().reset_index()
#     bot_counter['counter_Score'] = bot_counter.apply(lambda x: x['Player_4_gold']-x['Player_9_gold'], axis=1)
#     sup_counter = df[['Player_5_pick','Player_5_gold','Player_10_pick','Player_10_gold']].groupby(['Player_5_pick','Player_10_pick']).mean().reset_index()
#     sup_counter['counter_Score'] = sup_counter.apply(lambda x: x['Player_5_gold']-x['Player_10_gold'], axis=1)
#     return top_counter,jug_counter,mid_counter,bot_counter,sup_counter
def ChampCounterRelations(df):
    top_counter = df[['Player_1_pick','Player_1_gold','Player_6_pick','Player_6_gold']].groupby(['Player_1_pick','Player_6_pick']).sum().reset_index()
    top_counter['count'] = df[['Player_1_pick','Player_1_gold','Player_6_pick','Player_6_gold']].groupby(['Player_1_pick','Player_6_pick']).count().reset_index()['Player_1_gold']
    top1 = top_counter.loc[top_counter['Player_1_pick']< top_counter['Player_6_pick']]
    top2 = top_counter.loc[top_counter['Player_1_pick']> top_counter['Player_6_pick']]
    top2.columns = ['Player_6_pick','Player_1_pick','Player_6_gold','Player_1_gold','count']
    top_counter = pd.concat([top1,top2],ignore_index=True).groupby((['Player_1_pick','Player_6_pick'])).sum().reset_index().sort_values(['Player_1_pick','Player_6_pick'])
    top_counter['counter_Score'] = top_counter.apply(lambda x: (x['Player_1_gold']-x['Player_6_gold'])/x['count'], axis=1)
    
    jug_counter = df[['Player_2_pick','Player_2_gold','Player_7_pick','Player_7_gold']].groupby(['Player_2_pick','Player_7_pick']).sum().reset_index()
    jug_counter['count'] = df[['Player_2_pick','Player_2_gold','Player_7_pick','Player_7_gold']].groupby(['Player_2_pick','Player_7_pick']).count().reset_index()['Player_2_gold']
    jug1 = jug_counter.loc[jug_counter['Player_2_pick']< jug_counter['Player_7_pick']]
    jug2 = jug_counter.loc[jug_counter['Player_2_pick']> jug_counter['Player_7_pick']]
    jug2.columns = ['Player_7_pick','Player_2_pick','Player_7_gold','Player_2_gold','count']
    jug_counter = pd.concat([jug1,jug2],ignore_index=True).groupby((['Player_2_pick','Player_7_pick'])).sum().reset_index().sort_values(['Player_2_pick','Player_7_pick'])
    jug_counter['counter_Score'] = jug_counter.apply(lambda x: (x['Player_2_gold']-x['Player_7_gold'])/x['count'], axis=1)

    mid_counter = df[['Player_3_pick','Player_3_gold','Player_8_pick','Player_8_gold']].groupby(['Player_3_pick','Player_8_pick']).sum().reset_index()
    mid_counter['count'] = df[['Player_3_pick','Player_3_gold','Player_8_pick','Player_8_gold']].groupby(['Player_3_pick','Player_8_pick']).count().reset_index()['Player_3_gold']
    mid1 = mid_counter.loc[mid_counter['Player_3_pick']< mid_counter['Player_8_pick']]
    mid2 = mid_counter.loc[mid_counter['Player_3_pick']> mid_counter['Player_8_pick']]
    mid2.columns = ['Player_8_pick','Player_3_pick','Player_8_gold','Player_3_gold','count']
    mid_counter = pd.concat([mid1,mid2],ignore_index=True).groupby((['Player_3_pick','Player_8_pick'])).sum().reset_index().sort_values(['Player_3_pick','Player_8_pick'])
    mid_counter['counter_Score'] = mid_counter.apply(lambda x: (x['Player_3_gold']-x['Player_8_gold'])/x['count'], axis=1)

    bot_counter = df[['Player_4_pick','Player_4_gold','Player_9_pick','Player_9_gold']].groupby(['Player_4_pick','Player_9_pick']).sum().reset_index()
    bot_counter['count'] = df[['Player_4_pick','Player_4_gold','Player_9_pick','Player_9_gold']].groupby(['Player_4_pick','Player_9_pick']).count().reset_index()['Player_4_gold']
    bot1 = bot_counter.loc[bot_counter['Player_4_pick']< bot_counter['Player_9_pick']]
    bot2 = bot_counter.loc[bot_counter['Player_4_pick']> bot_counter['Player_9_pick']]
    bot2.columns = ['Player_9_pick','Player_4_pick','Player_9_gold','Player_4_gold','count']
    bot_counter = pd.concat([bot1,bot2],ignore_index=True).groupby((['Player_4_pick','Player_9_pick'])).sum().reset_index().sort_values(['Player_4_pick','Player_9_pick'])
    bot_counter['counter_Score'] = bot_counter.apply(lambda x: (x['Player_4_gold']-x['Player_9_gold'])/x['count'], axis=1)

    sup_counter = df[['Player_5_pick','Player_5_gold','Player_10_pick','Player_10_gold']].groupby(['Player_5_pick','Player_10_pick']).sum().reset_index()
    sup_counter['count'] = df[['Player_5_pick','Player_5_gold','Player_10_pick','Player_10_gold']].groupby(['Player_5_pick','Player_10_pick']).count().reset_index()['Player_5_gold']
    sup1 = sup_counter.loc[sup_counter['Player_5_pick']< sup_counter['Player_10_pick']]
    sup2 = sup_counter.loc[sup_counter['Player_5_pick']> sup_counter['Player_10_pick']]
    sup2.columns = ['Player_10_pick','Player_5_pick','Player_10_gold','Player_5_gold','count']
    sup_counter = pd.concat([sup1,sup2],ignore_index=True).groupby((['Player_5_pick','Player_10_pick'])).sum().reset_index().sort_values(['Player_5_pick','Player_10_pick'])
    sup_counter['counter_Score'] = sup_counter.apply(lambda x: (x['Player_5_gold']-x['Player_10_gold'])/x['count'], axis=1)

    return top_counter,jug_counter,mid_counter,bot_counter,sup_counter


# def MatchCounterSheet(x,top,jug,mid,bot,sup):
#     return float(
#         top.loc[(top['Player_1_pick']==x['Player_1_pick'])&(top['Player_6_pick']==x['Player_6_pick'])]['counter_Score']
#         ),float(
#         jug.loc[(jug['Player_2_pick']==x['Player_2_pick'])&(jug['Player_7_pick']==x['Player_7_pick'])]['counter_Score']    
#         ),float(
#         mid.loc[(mid['Player_3_pick']==x['Player_3_pick'])&(mid['Player_8_pick']==x['Player_8_pick'])]['counter_Score']    
#         ),float(
#         bot.loc[(bot['Player_4_pick']==x['Player_4_pick'])&(bot['Player_9_pick']==x['Player_9_pick'])]['counter_Score']    
#         ),float(
#         sup.loc[(sup['Player_5_pick']==x['Player_5_pick'])&(sup['Player_10_pick']==x['Player_10_pick'])]['counter_Score']    
#         )

# def CalulateCounterScore(df):
#     top = ChampCounterRelations(df)[0]
#     jug = ChampCounterRelations(df)[1]
#     mid = ChampCounterRelations(df)[2]
#     bot = ChampCounterRelations(df)[3]
#     sup = ChampCounterRelations(df)[4]
#     tmp = df.apply(MatchCounterSheet,args=(top,jug,mid,bot,sup),axis=1,result_type="expand")
#     tmp.columns=['topcounter','jugcounter','midcounter','botcounter','supcounter']
#     return tmp


def toCSV(top,jug,mid,bot,sup):
    N =4

    champ = pd.read_csv("../datasets/championId.csv")
    champ=champ[['id']]
    champ['id2'] = champ[['id']]
    champ = pd.MultiIndex.from_product([champ.id, champ.id2], names=['wepick','theypick']).to_frame().drop(['wepick','theypick'],axis=1).reset_index()
    champ = champ.loc[champ['wepick']<champ['theypick']].reset_index().drop('index',axis=1)
    
    df1 = champ.merge(top.loc[top['count']>N], 
    left_on=['wepick','theypick'], 
    right_on=['Player_1_pick','Player_6_pick'],
    how='left', 
    suffixes=(False, False)
    ).replace(np.nan,0)[['wepick','theypick','counter_Score']]
    df1.columns = ['T1','T2','T_counterScore']
    
    df2 = champ.merge(jug.loc[jug['count']>N], 
    left_on=['wepick','theypick'], 
    right_on=['Player_2_pick','Player_7_pick'],
    how='left', 
    suffixes=(False, False)
    ).replace(np.nan,0)[['wepick','theypick','counter_Score']]
    df2.columns = ['J1','J2','J_counterScore']
 
    df3 = champ.merge(mid.loc[mid['count']>N], 
    left_on=['wepick','theypick'], 
    right_on=['Player_3_pick','Player_8_pick'],
    how='left', 
    suffixes=(False, False)
    ).replace(np.nan,0)[['wepick','theypick','counter_Score']]
    df3.columns = ['M1','M2','M_counterScore']

    df4 = champ.merge(bot.loc[bot['count']>N], 
    left_on=['wepick','theypick'], 
    right_on=['Player_4_pick','Player_9_pick'],
    how='left', 
    suffixes=(False, False)
    ).replace(np.nan,0)[['wepick','theypick','counter_Score']]
    df4.columns = ['B1','B2','B_counterScore'] 

    df5 = champ.merge(sup.loc[sup['count']>N], 
    left_on=['wepick','theypick'], 
    right_on=['Player_5_pick','Player_10_pick'],
    how='left', 
    suffixes=(False, False)
    ).replace(np.nan,0)[['wepick','theypick','counter_Score']]
    df5.columns = ['U1','U2','U_counterScore']   

    output = df1.join(df2).join(df3).join(df4).join(df5)
    output = output[['T1','T2','T_counterScore','J_counterScore','M_counterScore','B_counterScore','U_counterScore']]
    output.columns=['id1','id2','T_counterScore','J_counterScore','M_counterScore','B_counterScore','U_counterScore']
    return output



def generateScoreSheet(df,path):
    top = ChampCounterRelations(df)[0]
    jug = ChampCounterRelations(df)[1]
    mid = ChampCounterRelations(df)[2]
    bot = ChampCounterRelations(df)[3]
    sup = ChampCounterRelations(df)[4]
    toCSV(top,jug,mid,bot,sup).to_csv(path, index=False)


generateScoreSheet(pre,"../Useful Features/1.counterScore_pre.csv")
generateScoreSheet(mid,"../Useful Features/1.counterScore_mid.csv")
generateScoreSheet(late,"../Useful Features/1.counterScore_late.csv")
