==> dataProcessingAlgorithms/step6.transformOriTraindataToSelectedFeatureData.ipynb <==
import GenerateInputDataForMLModel
from ChampionNameMapping import ChampionNameMapping
import pandas as pd


df_train_pre = pd.read_csv("../datasets/train_pre.csv")
df_train_mid = pd.read_csv("../datasets/train_mid.csv")
df_train_late = pd.read_csv("../datasets/train_late.csv")


# selected features
dt_pre = GenerateInputDataForMLModel.to_comparisonFeatureData(df_train_pre,"pre")
dt_pre.to_csv("../datasets/comparison_feature_data/train_pre.csv",index=False)


dt_mid = GenerateInputDataForMLModel.to_comparisonFeatureData(df_train_mid,"mid")
dt_mid.to_csv("../datasets/comparison_feature_data/train_mid.csv",index=False)

dt_late = GenerateInputDataForMLModel.to_comparisonFeatureData(df_train_late,"late")
dt_late.to_csv("../datasets/comparison_feature_data/train_late.csv",index=False)

# all features
all_pre = GenerateInputDataForMLModel.to_allFeatureData(df_train_pre,"pre")
all_pre.to_csv("../datasets/all_feature_data/train_pre.csv",index=False)

all_mid = GenerateInputDataForMLModel.to_allFeatureData(df_train_mid,"mid")
all_mid.to_csv("../datasets/all_feature_data/train_mid.csv",index=False)

all_late = GenerateInputDataForMLModel.to_allFeatureData(df_train_late,"late")
all_late.to_csv("../datasets/all_feature_data/train_late.csv",index=False)
