==> dataProcessingAlgorithms/step9.transformTestdataToSelectedFeatureDataGold.ipynb <==
import GenerateInputDataForMLModel
from ChampionNameMappingGold import ChampionNameMappingGold
import pandas as pd

df_test_pre = pd.read_csv("../datasets/val_pre.csv")
df_test_mid = pd.read_csv("../datasets/val_mid.csv")
df_test_late = pd.read_csv("../datasets/val_late.csv")

# comparison features
dt_pre = GenerateInputDataForMLModel.to_comparisonFeatureData_gold(df_test_pre,"pre")
dt_pre.to_csv("../datasets/comparison_feature_data/test_gold_pre.csv",index=False)


dt_mid = GenerateInputDataForMLModel.to_comparisonFeatureData_gold(df_test_mid,"mid")
dt_mid.to_csv("../datasets/comparison_feature_data/test_gold_mid.csv",index=False)

dt_late = GenerateInputDataForMLModel.to_comparisonFeatureData_gold(df_test_late,"late")
dt_late.to_csv("../datasets/comparison_feature_data/test_gold_late.csv",index=False)

# all features
all_pre = GenerateInputDataForMLModel.to_allFeatureData_gold(df_test_pre,"pre")
all_pre.to_csv("../datasets/all_feature_data/test_gold_pre.csv",index=False)

all_mid = GenerateInputDataForMLModel.to_allFeatureData_gold(df_test_mid,"mid")
all_mid.to_csv("../datasets/all_feature_data/test_gold_mid.csv",index=False)

all_late = GenerateInputDataForMLModel.to_allFeatureData_gold(df_test_late,"late")
all_late.to_csv("../datasets/all_feature_data/test_gold_late.csv",index=False)
